% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exdata}
\alias{exdata}
\title{Example Data}
\format{
## `exdata`
A data frame with 30 rows and 5 columns:
\describe{
  \item{BMI}{body mass index}
  \item{event}{death (0=alive, 1=dead)}
  \item{OS}{time of observation in months (minimum of event time and right censoring time)}
  \item{invasion}{depth of tumor invasion into stroma)}
  \item{LVSI}{lymphovascular space invasion (0=not seen, 1=present)}
}
}
\usage{
exdata
}
\description{
A small example dataset
}
\keyword{datasets}
