% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toy Data}
\format{
## `toydata`
A data frame with 537 rows and 5 columns:
\describe{
  \item{Age}{Subject Age}
  \item{BMI}{body mass index}
  \item{stage3}{binary (0=no, 1 = yes)}
  \item{Death}{death (0=alive, 1=dead)}
  \item{Death_surtime}{survival time}
}
}
\usage{
toydata
}
\description{
A small example dataset
}
\keyword{datasets}
